Automatically pull Authors, URL
Originally, Authors section was hard-coded. Instead, have added a for
loop to loop through each Authors in DESCRIPTION and, if role is
"aut", print Authors family name, given name, and with email that is
related to the new `ew.user_info` option, pull author's
git_user_profile as hyperlink. Prints results as markdown list.

Updated README to also pull DESCRIPTION URL instead of using
hard-coded repo URL, and Package field where only the package
name is listed.

The only place these are not used is in the badges but I suppose they
probably could/should be.